Support `stdin-placeholder` and `stdin-distinct` option
`stdin-placeholder` option receives a placeholder name that can be used
in `on-notified-cmd` to give the command the arguments that come from STDIN.

For example,

```
$ generator | conk -interval-sec 3 -on-notified-cmd '["echo", "{{__STDIN__}}"]' -stdin-placeholder '{{__STDIN__}}'
```

if the `generator` outputs `foo\nbar\nbuz\n` to STDOUT in 3 seconds, the conk
command interpolates those texts in `{{__STDIN__}}`, like
`["echo", "foo", "bar", "buz"].

And `stdin-distinct` option accepts the flag to instruct whether it makes the
arguments for `on-notified-cmd` that come from STDIN distinct, i.e. makes them unique.

```
$ generator | conk -interval-sec 3 -on-notified-cmd '["echo", "{{__STDIN__}}"]' -stdin-placeholder '{{__STDIN__}}' -stdin-distinct
```

in these commands, if the `generator` outputs `foo\nbar\nfoo\nfoo` to STDOUT in 3 seconds,
the conk command interpolates those texts in `{{__STDIN__}}` with unique
arguments, like `["echo", "foo", "bar"] because `foo` is duplicated.

Signed-off-by: moznion <moznion@mail.moznion.net>